+++ test/Qowaiv.Json.UnitTests/BsonMongoDBSerializeTest.cs
@@ -15,6 +15,7 @@ namespace Qowaiv.Json.UnitTests
             QowaivBsonConverter.RegisterType<SvoWithFromJson>();
             QowaivBsonConverter.RegisterType<SvoWithFromJsonClass>();
             QowaivBsonConverter.RegisterType<SvoWithFromJsonStringOnly>();
+            QowaivBsonConverter.RegisterType<Identifiers.Id<ForGeneric>>();
         }
 
         [Test, Ignore("For BSON a Date() is generated, that is default behaviour we don't want to interfere with.")]